requirements.txt
pandas==1.4.1
tqdm==4.64.0
openai==1.59.6
python-dotenv==0.21.0
streamlit==1.41.1
stqdm==0.0.5
plotly==5.7.0
wordcloud==1.9.3
matplotlib==3.9.2
scikit-learn==1.3.2